requirements.txt
apt-xapian-index==0.44
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
basemap==1.0.8
beautifulsoup4==4.0.2
cairocffi==0.6
CairoSVG==1.0.22
certifi==2016.9.26
cffi==1.8.3
chardet==2.0.1
command-not-found==0.2.44
cryptography==1.5.2
cssselect==0.9.2
cycler==0.10.0
enum34==1.1.6
ethtool==0.6
filelock==2.0.6
functools32==3.2.3.post2
GnuPGInterface==0.3.2
html5lib==1.0b8
httplib2==0.7.2
idna==2.1
ipaddress==1.0.17
keyring==0.9.2
Landscape-Client==13.7.3
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
lxml==3.6.4
matplotlib==2.0.0b4
numpy==1.11.2
oauth==1.0.1
oauthlib==2.0.0
PAM==0.4.2
paramiko==2.0.2
Paver==1.2.4
Pillow==2.3.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.4.1
pykml==0.1.0
pyOpenSSL==0.12
pyparsing==2.1.10
Pyphen==0.9.4
pyproj==1.9.5.1
pyserial==2.5
python-apt===0.8.3ubuntu7.2
python-crontab==2.1.1
python-dateutil==2.5.3
python-debian===0.1.21ubuntu1
pytz==2016.7
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
screenutils==0.0.1.6.1
simplejson==2.3.2
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
tinycss==0.4
tornado==3.2
tornado-redis==2.4.18
Twisted-Core==11.1.0
twython==3.4.0
ufw==0.31.1.post1
virtualenv==15.0.3
wadllib==1.3.0
WeasyPrint==0.31
webencodings==0.5
websocket-client==0.40.0
zope.interface==3.6.1
